requirements.txt
requests==2.23.0
splunk-sdk==1.6.12
PyYAML==5.3.1